+++ include/tmv/TMV_ProdXV.h
@@ -70,7 +70,7 @@ namespace tmv {
         TMV_INLINE const Scaling<ix,T>& getX() const { return x; }
         TMV_INLINE const V& getV() const { return v; }
 
-        TMV_INLINE size_t size() const { return v.size(); }
+        TMV_INLINE int size() const { return v.size(); }
         value_type cref(int i) const 
         { return  x*v.cref(i); }
