+++ include/tmv/TMV_QuotVM.h
@@ -221,7 +221,7 @@ namespace tmv {
         TMV_INLINE const V1& getV() const { return v1; }
         TMV_INLINE const M2& getM() const { return m2; }
 
-        TMV_INLINE size_t size() const { return m2.rowsize(); }
+        TMV_INLINE int size() const { return m2.rowsize(); }
 
         template <class V3>
         TMV_INLINE_ND void assignTo(BaseVector_Mutable<V3>& v3) const
@@ -288,7 +288,7 @@ namespace tmv {
         TMV_INLINE const V1& getV() const { return v1; }
         TMV_INLINE const M2& getM() const { return m2; }
 
-        TMV_INLINE size_t size() const { return m2.colsize(); }
+        TMV_INLINE int size() const { return m2.colsize(); }
 
         template <class V3>
         TMV_INLINE_ND void assignTo(BaseVector_Mutable<V3>& v3) const
